Update construal_level_task.py
Removed "is_first = False" from line 55: whenever the HLC task was first within a run, it would not show the "Calibrating Scanner" page and auto-started the task without waiting for the trigger pulse from the scanner. It also showed the "waiting for the researcher" page following the task, forcing the researcher to hit the spacebar to manually advance to the second task in the run. By removing this one line of code, both problems have been resolved.